# struct: store members as starlark.StringDict for faster access