@emcfarlane 2022-03-07
Closing for https://github.com/google/starlark-go/pull/403